requirements.txt
web3
discord.py
python-dotenv
requests